llm_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch>=2.3.0
transformers>=4.41.0
accelerate>=0.30.0
pillow>=10.3.0
pydantic>=2.7.0
bitsandbytes>=0.43.0
python-multipart>=0.0.9
qwen-vl-utils>=0.0.1
